+++ lua/plugins/practicalli.lua
@@ -123,6 +123,17 @@ return {
       }
     end,
   },
+  -- Switch between src and test file
+  -- TODO: PR #67 raised on rgroli/other.nvim
+  {
+    "practicalli-johnny/other-nvim-fork",
+    branch = "clojure",
+    ft = { "clojure" },
+    main = "other-nvim",
+    opts = {
+      mappings = { "clojure" },
+    },
+  },
   -- ------------------------------------------
 
   -- ------------------------------------------